+++ mpv.net/DynamicGUI/DynamicGUI.cs
@@ -29,7 +29,6 @@ namespace DynamicGUI
                     baseSetting = optionSetting;
                     optionSetting.Default = setting["default"];
                     optionSetting.Value = optionSetting.Default;
-                    optionSetting.StartValue = optionSetting.Default;
 
                     foreach (TomlTable option in setting["options"])
                     {
@@ -70,7 +69,6 @@ namespace DynamicGUI
     {
         public string Name { get; set; }
         public string Value { get; set; }
-        public string StartValue { get; set; }
         public string Help { get; set; }
         public string Default { get; set; }
         public string URL { get; set; }